+++ Models/Nokia.cs
@@ -2,7 +2,7 @@ namespace Models;
 
 public class Nokia : Smartphone{
     // Construção da classe
-    public Nokia(string numero, string modelo, string imei, int memoriaTotal) : base(numero, modelo, imei, memoriaTotal) {
+    public Nokia(string numero, string modelo, string imei, float memoriaTotal) : base(numero, modelo, imei, memoriaTotal) {
     }
 
     // Implementação do método abstrato
